Add time to log name; Skip log prompt during "force" (#24)
When I was running `helm uninstall mmai -n cattle-system; ./cleanup.sh
force` I realized that the cleanup command can hang on the logfile
prompt even if force is provided. Changes:

1. change the default logname so that it has a date at the end.
2. if "force" is provided and there's already a file with the default
name, we append ".2"; if a "file.2" already exists, we take the number
and increment it by one.
3. if "force" is not provided, ask for the logname after the continue
(y/n) prompt.